feat(keys): Report keysChangedAt in assertions and key ids.
The "durable sync" project wants to bucket storage of server-side
sync data based on the id of the key used to encrypt it, including
a timestamp component so that it can easily determine which data
is newest. Our OAuth scoped key ids include a leading timestamp
designed for just this purpose.

Up until now we have been using `verifierSetAt` for that timestamp.
This works but means that we generate a new kid whenever the user
changes their password, even if that did not involve resetting the
master encryption key `kB`. This means that durable sync will give
users a new storage bucket, in  response to a password change,
discarding the user's previous sync data even though their key had
not changed.

Instead, we need to explicitly track the timestamp at which key
material actually changed, and report that in both OAuth key ids
and in BrowserID assertions so that all Sync clients see a consistent
timestamp.  We already have a `keysChangedAt` database column from
a previous migration, but haven't been making any use of it, and it
currently defaults to the value of `verifierSetAt`.

With this commit, we start using `keysChangedAt` for the timestamp
component of OAuth scoped key ids, and start reporting it in
BrowserID assertions. For now the reported value will always be
the same as `verifierSetAt`, which will let us get the new logic
into production without changing what timestamps are observed by
the sync service.

A follow-up commit will be required to actually start tracking
`keysChangedAt` separately from `verifierSetAt`.

Connects to #490.