chore(metrics): de-categorise content server perf events
Historically we divided content server performance events into two
categories, "auth" and "other". That was useful for viewing agggregate
performance at the start of our funnels, but got in the way of other
analysis. This change ditches those categories and just uses the raw
view names instead. If we need to categorise events, we can do that in
bespoke queries in Redash.