fix(authorization): Fix timezone issues when trading a code for a token
Times written to the OAuth database are in UTC time. Times read out
do not have timezone information and are assumed to be in local time.

In London for this time of the year, that means we are 1 hour ahead
of UTC, and all dates read from the DB are one hour in the past.

Specify a `timezone` when configuring the MySQL connection, default
to `UTC` which is the timezone of the MySQL database.

fixes #2626