fix(oauth): Revoke access tokens when revoking a refresh token
All access tokens for a client_id, uid pair are revoked whenever
revoking a refresh token to avoid phantom entries in the
user's devices & apps list.

Before this PR, the problem was revoking a refresh token would
only revoke the refresh token and none of the access tokens created
from it. This led to a user revoking a refresh token, but then
the access tokens created from it showing up in the user's devices
& apps list, and needing to revoke those separately.

Because we don't associate access tokens and refresh tokens, we can't
delete just the access token associated with the refresh token. Instead,
this PR uses a blunt instrument and revokes *all* the access tokens
for the user/client_id. If the user has multiple active refresh tokens,
then the clients will use those refresh tokens to get new access
tokens for the ones that were revoked prematurely.

issue #1249