refactor(db): Remove the memory DB in the Auth server
The original idea of the memory DB was so developers
didn't need to install MySQL on their local machines.

Now that we run MySQL in a container, running MySQL
is no longer the burden it once was. This old memory
DB is just crufty code that has to be maintained and
often slows us down.

Goes along with #3062 where the memory DB was removed
for the OAuth server.