fix(test): Fix the OAuth TOTP required tests
The beforeEach seemed to do too much to me, like it was testing
the signup flow, and that slowed the actual tests down because
too much was happening. So, I created a separate signup test
that tests that flow.

I could not solidly pinpoint a reason why the tests sometimes
failed and sometimes didn't, but my guess is because generateTotpCode
was called before TOTP was even confirmed. Because remote tests
can take a significant amount of time (and in fact I saw this test
time out several times while running against latest), the TOTP
code may have been expired by time it was entered into the form.

The major change here is to generate the TOTP code when the form
should be filled out, not before.

Other changes are to make heavier use of helpers to reduce
boilerplate signin/signup code.

fixes #3121