fix(browserid): Let clients specify service=${client_id} in /certificate/sign
Due to some issues with processing the high volume of OAuth token-creation
events, we may need to have desktop Firefox use BrowserID assertions to
create its OAuth tokens, meaning it may want to call /certificate/sign even
if the `sync` service is not enabled.

This commit allows values other than `sync` in the `service` query param
for /certificate/sign, so that we won't get bogus metrics from browsers
that are calling this endpoint but are not connected to sync. For more
context see https://bugzilla.mozilla.org/show_bug.cgi?id=1591312.