fix(payments): include time fields in amplitude events
Specifically, include the `perfStartTime`, `startTime`, and `flushTime`
in each amplitude event emitted by the front-end.

More details:

The content-server amplitude events are also flow events, meaning
the `perfStartTime` and other times needed to reconstruct the event time
on the server are already available on the event object. The arithmetic
used to calculate `event.time` was copied over to the payments server
post-metrics route handler code without verifying that the needed data
was available on the amplitude events emitted by the front-end.

Unlike content server, the payments server amplitude events are
generated by a redux middleware that observes certain actions dispatched
while the user interacts with the payments pages, which is separate from
how payments flow events are generated and processed. This means that the
necessary time values need to be explicitly added to each amplitude event
before the event is sent down to the server.

I verified this patch fixed the issue by attaching the node debugger to
a local payments server process, but you could also look in the payments
server logs for an amplitude event logged with a `time` property. One
such event (the engage event) is emitted as soon as a character is typed
anywhere in the subscription payment form.

Hopefully, this really fixes #3050.