fix(email): Ensure the action buttons are translated
The action button text was not being translated. The
action button text was wrapped in `gettext`, but `gettext`
on its own is only for string extraction. To translate,
the text needs to be run through `translator.gettext`
rather than just gettext.

I opted to do this in the `localize` method so
that translation could be centralized.

fixes #3258