feat(metrics): Allow RPs to submit a generic 'engage' event to amplitude
With this change, relying parties can submit an 'engage' ping to the
get-metrics-flow endpoint as a way to record that a particular FxA user
has interacted in their service in some meaningful way. Only one event is
possible per RP in this approach, and the meaning of the 'engage' event
has to be defined out of band.

This approach is a bit of a hack, since the 'engage' event for Monitor
isn't really a flow event, but a ping fired by the Monitor server when a
user visits the site after clicking a breach alert email link. Also, the
raw FxA uid is included in the request; longer term, we should use the
anonymized uid equivalent being developed as part of ecosystem telemetry
work.

Fixes #2743; see the bug for additional discussion. See also the data
collection review bug, https://bugzil.la/1594388.