fix(tests): update a support form functional test
PR #2824 updated payments server to redirect back to settings if a user
does not have any subscriptions.  Subsequently the relevant functional
test was updated in PR #3147.  However, there was a timing issue in the
test.  Since the payments server will redirect at some point during its
rendering process, there may not be enough time to do the assertions in
the test.

This patch updates the test to not perform any assertions on the
subscriptions management page.

Part of #3364