chore(test): Avoid timing issues in sign_in_totp tests
The reset password tests did not wait for the CONFIRM_RESET_PASSWORD.HEADER element
before opening the links from the emails. I haven't seen this be a problem
in practice, but am opening the PR as a belts and suspenders move.

Also removes the listenForFxCommands, it's not necessary for the tests
to pass since it is only used for fx_desktop_v1 and fx_ios_v1.