refactor(test): Remove Firefox Desktop 57 and below in the tests
FxA officially only supports Fx desktop v60 and above, yet we still
have lots of code that checks the behavior of Fx 57 and below.

The major difference between Fx 57 and Fx 58 is the removal of
about:accounts. In Fx 57 and below, after verifying a login, about:accounts
was expected to take over and FxA was not supposed to do a screen
transition.

Starting with Fx 58, about:accounts no longer existed and FxA
needed to show a "you have finished the flow" screen, which allowed
us to display /connect_another_device.

Since the only major change was this transition, which if does
occur on Fx 57 and below isn't the biggest deal, I removed
support for preventing the transition and just transition
everywhere. This simplifies broker logic, it simplifies test
logic, and means we have fewer things to think about.

Also, since about:accounts was removed in Fx 58, I removed the
logic for "isAboutAccounts" within the code.

Not attached ot an issue.