refactor(test): Simplify test setup when WebChannel responses are expected
The `automatedBrowser` query param is needed whenever opening a page that
hooks up WebChannel listeners. The query param causes FxA to introduce a slight
delay before firing the fxaccounts:fxa_status WebChannel message, the delay allows
the tests to hook up listeners that intercept and respond to the message. Without
the delay, the fxaccounts:fxa_status message fires too early and no response
occurs, causing the tests to stall.

Up to this point, developers had to remember to add the `automatedBrowser`
query parameter. This ensures it's added automatically if `openPage` is
called with `webChannelResponses`, relieving the dev from remembering
the need to do this.

Not attached to an issue.