refactor(experiments): Simplify how experiments and metrics are created
Previous to this, if a user was part of an experiment, we had to
manually call `this.createExperiment` within a view to ensure the
metrics are reported.

I added a new method, `this.chooseExperimentGroup` to the ExperimentInterface
that automatically calls `this.createExperiment` for users that are
part of the experiment.

All the places where `getExperimentGroup` then `createExperiment` were
called have been replaced with a new call to `chooseExperimentGroup`,
simplifying setup and reducing the number of tests that are needed.

Not attached to an issue, noticed this while writing docs on how
to add a new experiment:

https://mozilla.github.io/ecosystem-platform/docs/fxa-engineering/fxa-content-server-ab-testing