chore(metrics): remove flow perf metrics from Payments
This patch removes flow based performance metrics on Payments server.  The
data from the logs are not present in Redshift due to the lack of a log
pipeline to ingest and process the events.  Furthermore, there was a bug
with the timing values.

This will be followed up with a switch to statsd, so that performance
metrics will be separated from flow events.