refactor(tests): Expose all TestHelpers within FunctionalHelpers.
While writing the functional test deep dive, I found myself thinking
it odd I had to explain that two files had to be included to create
a user. So I decided to make it so one file needed to be included
by exposing all TestHelper methods in FunctionalHelpers.

Not attached to an issue.

ref https://github.com/mozilla/ecosystem-platform/pull/47