fix(oauth): Do not store raw access tokens in Redis.
The MySQL access-token storage deliberately does not write the raw
access token to disk, and instead only writes a hash. This allows
us to look up a given token but prevents a db compromise from immediately
leaking all tokens in useable form.

The new Redis access-token storage currently includes the raw token
in its JSON serialization. We should never need this value again in
practice, so I suggest we avoid storing it for all the same reasons
we don't store it in MySQL. This commit changes the serialization
logic to persist `tokenId` instead, while continuing to support data
with the old field that may already be in redis.

Relatedly, the value returned when looking up an access token could
differ depending on whether it came from Redis (in which case the
`token` field would be the raw token) or from MySQL (in which case the
`token` field would be the hashed token id). I've included a bit of
a refactor of how the `AccessToken` class is used, so that these
differences are encapsulated in the DB layer and calling code will
only ever see instances of the `AccessToken` class