feat(admin-panel): Build out initial UI with nav
This commit sets up routes and introduces a lot of new elements and styles for the beginning of the admin panel email block clearing feature. Components were created with reuse in mind for other packages but styling conventions (including class names, nesting, placement, and use of CSS variables) and components may change as the panel is built out. This commit does not introduce any state just yet.

fixes #4204