fix(access-tokens): prune the access token index on set
The access token index can grow too big since the only
chance to prune until now was when the full list was
requested, which doesn't happen often.

This change adds a limit config and prunes when a new
token is added if the index is over the limit. It makes
no effort to check expiry so in the worst case up to 10%
of the limit active tokens will be deindexed, but still active.

To limit the size of the initial pruning no more than limit
tokens will be deindexed per call.

The index set will now be set to expire after the configured
max ttl for access tokens. Each call extends the expiration.