feat(email): add rel="noopener noreferrer" to any email links that have target="_blank"
This attempts to address a potential attack vector where clicking a link can redirect you to a malicious website. More info: https://dev.to/dhilipkmr/why-should-you-use-noopener-beware-of-security-flaws-3i57

This sort of attack is less likely to occur in email, but is technically still possible in a web client. More info: https://stackoverflow.com/a/51194051/717633