fix(settings): wrap overflowing settings group headers so they don't break layout
In English none of our settings titles are long enough to bump up against the right side button. Lucky us. However in localized versions of the settings panel these titles can get quite long. Currently long strings push the content of the element out, effectively breaking the settings layout.

This fix allows the text to wrap to a new line. To accomodate this I've changed the explicit height of the settings row to a minimum height, and added vertical padding equivelent to the space between the right-side button and the top/bottom of the container.