feat(metrics): validate Amplitude events with a JSON schema
This patch adds a JSON schema to fxa-shared and uses it to validate the
Amplitude events in the auth, content, payments servers.

At the moment when an Amplitude event fails validation against this
initial schema, it is still emitted.  This will be changed in a future
commit once the degree of confidence on the schema is higher.