feat(metrics): log raw Amplitude events in payments server
This patch logs a "raw" Amplitude event in payments server to stdout
along with enough context for another service to transform the raw event
into an Amplitude event identical to one that's emitted by
fxa-amplitude-send.

It also removes a unused parameter from the exported amplitude function;
that change is not directly related to the raw event changes.