fix(build): change fxa-js-client resolution in content-server
Due to fxa-js-client being a package.json dependency in content-server
and lerna/renovate trying to be "helpful" there was inconsistency
with how the package was resolved based on the environment. In Docker
and CI it loaded from npm but in local dev via an npm link to the
monorepo source. This change just stops using npm resolution at all.