refactor(config): replace 127.0.0.1 with localhost
why, after all these years?

because there's a few annoying cases where the loopback
makes dev harder. When you try to test a flow from a not
local machine, like a tv, mobile device, or a saucelabs
proxy session. With localhost it's easy enough to forward
but loopback not so much

enough is enough