fix(content-server): fixed build-production after #4763
Since #4763 introduced async/await the uglifyjs minimizer
fails. The built in webpack minimizer (Terser) does handle
it and performs about equally well.

The other issue is whether we transpile async/await for
older browsers, hence the addition of babel-loader to .ts
files. I opted for the safer option of doing the transpile
for now. We can defer that decision until later.