fix(npm): fixed npm install in content-server etc
fixes #5086

This is an interim fix to the package-locks for
internal dependencies, i.e. "file:.." ones. There
appears to be at least one bug in npm in parsing
a package-lock with more than one level of depth.

I couldn't find a way to make both lerna and npm
happy at the same time. I think the real fix will
be a different approach to handling internal deps
and their build order resolution.