fix(l10n): temporary (hopefully) fix for l10n CI build
The l10n repo build installs things differently from
this repo. The `npm ci` on fxa-auth-server is "broken"
with how deps are currently laid out. We really need
a better way to manage interpackage dependencies, but
that's a more involved issue.

In the meantime adding the @types from fxa-auth-server
to the fxa-shared tsconfig fixes the build error.

ref: https://github.com/mozilla/fxa-content-server-l10n/blob/8a5515c75395169b2f1badbdc03f36f5dd5fbd99/scripts/extract-and-pull-request.sh