feat(oauth): support jwt access tokens for Sync
This change enables Firefox Sync to transition
from BrowserId based authentication to OAuth.

In order to do so scalably we begin allowing
short lived JWT access tokens to be verified
directly without storing them on the server.
Since doing so will benefit other parts of
the system, like profile-server we enable
this globally based on the
JWT_ACCESS_TOKEN_CLIENT_IDS env variable.

This also adds the fxa-generation field to
the JWT to help Sync keep track of changes.