feat(aet): add ecosystem_anon_id field to auth server core profile data
The ecosystem_anon_id work for AET (Jira epic FXA-880) started from the
API used by desktop code, in order to unblock that work, which can be
done in parallel. This new identifier is eventually going to be part
of the core profile fields stored in the auth server DB, which means
there will be changes in mysql, auth-db, auth-server, and profile-server.

The initial no-op profile-server API implementation landed as
https://github.com/mozilla/fxa/pull/5364.

This change moves one step further down the stack, updating the
profile-server API handler from a no-op to an actual call into the
_core_profile, and adding a deceptively simple change to the auth server
that allows an ecosystemAnonId field to be passed through from auth-db.

Future commits will integrate this new field into the auth-db and
eventually into mysql.