fix(docker): removed need for fxa-shared postinstall script
The postinstall of fxa-shared was a temporary hack to prevent
auth-server from failing during it's build step. This unfortunately
caused the build-builder.sh script to fail when it tried to
only include the non-dev dependencies from the docker image.

The fix was to make fxa-shared a composite ts lib and add
a reference to it in fxa-auth-server's tsconfig. This required
using a 'path' instead of direct fs access in order to work
with both ts-node for dev and tsc for prod.