fix(oauth): Allow token requests to specify arbitrarily large TTL values
If an access token is requested with a longer TTL than the configured
maximum, the request is rejected as invalid. This excessively strict
validation prevents us from lowering the token TTL without also updating
every RP that uses tokens, and also does not follow the behavior
described in our API docs.

This change relaxes validation of the `ttl` parameter in token-issuing
routes, allowing arbitrarily large TTL values to be requested, while the
route handler behavior will now return the smaller of the requested TTL
and the maximum TTL set in server configuration.

Fixes #5660.