fix(build): added postinstall to fxa-shared
In dev we want to be able to go from a `yarn install` to
running code but we also want to keep typescript as a
devDependency for production builds where we have an
explicit build phase then --production install. To keep the
production install from failing on the postinstall we need
it to succeed even if tsc is missing, hence the `|| true`.