fix(l10n): fixed currency localization for "other than US" locales
For en locales that don't have their own ftl file this change
sets their FluentResource to be that of en-GB, which does have
an ftl file in order for the currency to be rendered as 'US$x.xx'
instead of the US fallback of '$x.xx' which is ambiguous in which
currency is meant.

These locales are currently hard coded in AppLocalizationProvider
because other services fail when they can't load l10n files that
are present in supportedLanguages.json but not in the l10n repo.