chore(build): build fxa-email-service less often
We're building email-service on train tags more often
than we need to and doing so adds a good chunk of time to the
build process.

We've already got some logic in hash-source.sh to determine
when a build should happen. This change adjusts the criteria
to exclude the files that change when a tag is done even when
there's no meaningful changes.