refactor(rescripts): include all of fxa as valid paths in rescripts
Because I was about to add fxa-auth-client to the additionalJSIpmorts but ran into problems. Instead of trying to include the minimum set of paths to add to the webpack rule this just lets all of fxa in.

Also, since we started using `workspace:*` in package.json we no longer need the aliases or jest module mappings!