refactor(settings): removed AccountDataHOC
Because our assumptions about the data from gql have changed, we no longer need to worry about parts of the Account type needing a verified session, we can simplify the structure a bit.

In the future we can use `useAccount` in components to get the data they need instead of pushing down the `account` object.

This also introduces the `models` folder for organizing our types.

For testing components that `useAccount()` there's the `MockedCache` component.