feat(settings): display alert messages from login in new settings
Some actions in fxa-content-server end up displaying an alert message on the settings page. We want to be able to display those messages across the "app" boundary between the content-server front-end and new settings.

To do this we store the message in localStorage and whichever app gets to it first will display it and remove it from storage.