fix(content server): make safari save the correct fields in keychain
Because:

- Safari does not effectively save the email you enter with the KeyChain login item. It instead grabs the age.

This commit:

- Attempts to direct Safari to the appropriate fields.

My current theory is that it has something to do with the field being `disabled` or set to `display: none`. From my tests, using a `name` appeared to direct Safari to the correct fields, and this still appears to work in other browsers. Also, because we don't submit the form in the traditional way, the `name` on inputs has no impact on the data going in.