refactor(settings): limit queried data for reloads
Because:
 - querying everything for an account when reloading the state of
   recovery key, two step auth, or connected services incur unnecessary
   costs

This commit:
 - use reload queries that only query gql-api for what they need

The most impactful change is actually the 'keyFields' configuration for
Account in the InMemoryCache's 'typePolicies'.  It allows the query
results from the reloads to be automagically merged into the cache.
(The value of 'keyFields' is set to '[]', making the cached account a
singleton.)

However, setting 'keyFields' affected the `cache.modify` calls to update
the cached account.  The account passed into the field modifier function
is now a Reference, not a plain object, thus breaking the "clone object
then merge in updated properties" pattern.

Setting the proper `__typename` property is now also vital, at least for
Account.  Without it the cache will not able to merge updated properties
into the cached account.  (gql-api already include `__typename` on its
results.)