fix(oauth): fixed oauth -> auth error conversion after refactoring
In anticipation of combining the auth/oauth error.js source files I prematurely removed the pre-existing logic to convert oauth errors into auth errors for certain routes. This restores that logic in the `appErrorFromOauthError` function and calls it as part of `AppError.translate` where appropriate.

I also renamed the AppError 'type' in oauth/error.js to OauthError to make them easier to differentiate.