fix(settings): ensure modal callbacks adhere to the defined type
The Modal and ModalVerifySession components were incidentally calling the onDismiss callback with the dom event argument from their originating element. This happened to break the ConnectedServices component which was using a fn that accepted optional parameters as the callback.

This commit ensures the callbacks are called with no arguments.

fixes #7509