task(settings): upload avatar directly to profile server
Issues #7536 #7560 #7556 are related to bugs in the gql upload implementation. Since the work required to fix them is unknown and there is an approaching milestone for new settings, it's more time effective to implement the upload directly to profile server. We can continue developing the gql upload and switch over when it's ready for production.