fix(settings): show notification on task completion
When tasks in a sub-page are completed, the user is redirected
back to the parent settings page. In this commit, we use the
alertTextExternal reactive variable to hand off the task of displaying
notification text to the parent page.

This added behavior proved difficult to unit test for two of the
components, and that work has been deferred to #7596, which shouldn't
block release.

Fixes #7030.