task(fxa-settings): remove unused file
This file must have been readded in a borked rebase or something during
the avatar work, I remember having to address a couple of different
conflicts at the time. I just noticed this while trying to link to some
code for an avatar bug. This is embarresing but I guess better late then
never. \_(ツ)_/