feat(emails): add email for payment method cancellation
Because:
 - we should let a customer when their payment method for active
   subscriptions has a problem

This commit:
 - send an email on a billing agreement cancelled event from PayPal

This commit also changes the declarative email tests from a Map to an
array.  A previous patch (mistakenly) treated the tests as an array, and
this patch changes that list of tests to match.  The change to an array
allows multiple tests of the same mailer method, which we need in this
patch.