feat(profile): create a monogram default avatar
When an account has no avatar set we attempt to create a default avatar using the first letter of either display name or email address. These avatars use theme colors within the Firefox native UI and a neutral grey in web content and are served by the profile server.

To test with theme colors in Firefox change `svg.context-properties.content.enabled` to `true` in `about:config`.