feat(auth): convert remaining auth-server db use to direct db access
This eliminates the runtime dependency of auth-server on auth-db-mysql. It's no longer required to run the auth-db-mysql process in production for the auth-server. This change doesn't completely remove the need for auth-db-mysql, email-service still needs it, and in dev a portion of it is required to run migrations, but it should now be considered deprecated and we should continue working to remove it completely.