fix(pm2): resolve nest cli path in pm2 configs
There's a few cases where the 'nest' command might not be in the PATH when pm2 tries to run the script. One example is starting a single service with `yarn workspace fxa-graphql-api start`. If we resolve the path we don't need to worry about what environment the command gets run in.