fix(fxa-auth-server): handle prorations during invoice detail extraction
Because:
The wrong email is sent to the user when upgrading between different
products since the invoice detail extraction only pulls from the first
line item in the invoice. In the case of a prorated invoice, the first
line item will always be the proration, which in the case of
cross-product upgrades will be the wrong plan.

This commit:
Filters all prorated line items for invoice detail extraction.

Closes FXA-5733