fix(payments-server): return inactive plans
Because:
The plans endpoint previously did not return inactive plans. This caused
issues with users who are subscribed to a plan which was then
subsequently marked inactive.

This commit:
Returns inactive plans from the plans endpoint, along with the active
flag. The payments frontend now checks to see if a plan is inactive
before allowing the user to proceed through the payment flow.

Closes FXA-5599