fix(fxa-admin-server): throw error when iap sub has no matching stripe plan
Because:

Due to some typecasting and improper typing, the admin server would try
to return null for some properties which were declared as non-null in
the schema if there existed an IAP subscription with no matching stripe
plan. This would cause GraphQL to throw a "cannot return null for
non-nullable field" error.

This commit:

Throws an error when the user has an IAP subscription with no matching
stripe plan.
There is further work to do in this area of the codebase, however. The
underlying issue at play here is the improper typing and typecasting of
many of the fields assembled to serve these requests. A followup ticket
will be made to solve these issues.

Closes FXA-5820