fix(fxa-settings): wrap overflowing words to prevent overlapping text Because:
* Currently, very long words (such as in German) are following standard
  rendering behavior in breaking out past their element containers in
the Security panel of the Settings page. This is considered a preferred
default because it prevents data loss (for example, from hiding the
overflow.) It is undesirable in our situation because the words are currently overlapping
with, and obscuring, other elements.

This commit:

* Applies "break-words", the tailwind class which corresponds to
  "overflow-wrap: break-word", to the h3 label in all UnitRow
components. This means that if a label (even a single-word label)
exceeds the horizontal bounds of its container, it will wrap, keeping
all text within the element visible and non-overlapping.

Closes #https://mozilla-hub.atlassian.net/browse/FXA-6089

Co-authored-by: vpomerleau <v.pomerleau@gmail.com>