storybook(fxa-settings): add pair/auth/totp view to storybook Because:
* We're recreating the remaining backbone views in React in storybook as
  a preliminary step to replacing the backbone views. This recreates
pair/auth/totp. In backbone, this view uses the same template as
signin_totp_code. The two main differences are 1. the action which
happens on submit, and 2. this view hides the alternatives to totp. I've
opted to create a new component since so many of our components have
ended up being a bit unwieldy due to reuse, but I'm on the fence as to
whether that was the right choice. Let me know your opinions! Happy to
change it.

This commit:

* Copies the signin_totp_code code with the needed changes. Since we
  reuse strings, it doesn't involve any new strings. It does involve
slightly different metrics.

Closes #FXA-6751