chore(deps): upgrade react-scripts and webpack to v5
Because:
 - we want to upgrade our React/CRA packages to react-scripts and
   webpack v5

This commit:
 - upgrades react-scripts and webpack to v5 for Settings, Payments, and
   Admin Panel
   - upgrades other dependencies in order to make the react-scripts and
     webpack upgrades work
 - hybridizes fxa-auth-client and fxa-shared into dual module format
   packages

The most consequential change in this patch is Webpack 5's support of
"exports" in package.json.  Majority of the changes are a result of
that.  With the support of "exports", it means that, without any other
changes, Webpack 4 and 5 could be transpiling and bundling different
files. E.g. Webpack 4 transpiling the original source while Webpack 5
loads the file in "exports" that's been transpiled by tsc.