task(settings): Support CDN in fxa-settings
Because:
- Settings is not being served from a CDN

This Commit:
- Configures webpack so that it can build assets that will work with a CDN
- Updates docker build to produce static bundle for stage cdn
- Updates docker build to produce static bundle for prod cdn
- Makes a couple changes to target these new build assets in content server
- Uses the full hash in bundle names to reduce the chance of conflict

Note, that we currently do a string replacement on index.html to inject some config and a flow id. If we didn't do this, we could serve this as a SPA directly off the CDN which would be faster. This should be a future consideration.

To test what this would look like once deployed on stage / production, override the following in local.json config.

  "static_directory": "dist"
  "page_template_subdirectory": "dist"
  "proxy_settings": false